xml serialization is now using separated module
