+++ source/deps.ts
@@ -1,6 +1,9 @@
+// Third party 
 import marked from "https://raw.githubusercontent.com/olaven/marked/strict-types/main.ts";
 export {marked as markdown_to_html}
+export { serialize, Tag } from "https://raw.githubusercontent.com/olaven/serialize-xml/v0.1.0/mod.ts"
 
+//standard Deno lib 
 export { decode } from "https://deno.land/std/strings/decode.ts";
 export { encode } from "https://deno.land/std/strings/encode.ts";
 export { green, red, yellow, bold } from "https://deno.land/std/fmt/mod.ts";
