+++ source/rss/converter.ts
@@ -1,4 +1,4 @@
-import { Tag } from "./serialize/mod.ts"
+import { Tag } from "../deps.ts"
 import { Post, Collection } from "../blog/collection.ts"
 import { get_rss } from "./rss.ts"
 import { Blog, Item } from "./types.ts" //TODO: Decouple! Too many imports
