+++ source/rss/types.ts
@@ -1,4 +1,4 @@
-import { Tag } from "./serialize/mod.ts"
+import { Tag } from "../deps.ts"
 import { Collection } from "../blog/collection.ts"
 
 //TODO: this interface is only used by rss. Should it be generalized for more concepts in this program?
